Addded memory for primitivePolygonIndices

+++ modules/wkt/src/lib/parse-wkb.ts
@@ -318,7 +318,7 @@ function concatenateBinaryPolygonGeometries(
     type: 'Polygon',
     positions: {value: concatenatedPositions, size: dimension},
     polygonIndices: {value: new Uint32Array(polygonIndices), size: 1},
-    primitivePolygonIndices: {value: new Uint16Array(primitivePolygonIndices), size: 1}
+    primitivePolygonIndices: {value: new Uint32Array(primitivePolygonIndices), size: 1}
   };
 }
